chore: move config gen to different file
